pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asrai-mcp"
version = "0.4.4"
description = "Asrai crypto analysis MCP server — pay-per-use via x402 on Base"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.6.0",
    "x402>=0.2.0,<1.0.0",
    "eth-account>=0.13.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
sse = [
    "starlette>=0.36.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
asrai-mcp = "asrai_mcp.server:main"
asrai-mcp-server = "asrai_mcp.sse_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/asrai_mcp"]
